Created a new file for events processing
